pyproject.toml
[project]
name = "vector-brain"
version = "0.1.0"
description = ""
authors = [
    {name = "Suender Oliveira", email = "suender@live.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-core (>=0.3.63,<0.4.0)",
    "streamlit (>=1.45.1,<2.0.0)"

]



[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.9.0",
    "mypy>=1.5.0",
    "flake8>=6.0.0",
    "pytest-cov>=4.1.0",
    "ipython>=8.12.0",
    "pre-commit>=3.3.3",
]

[project.scripts]
vector-brain = "vector_brain_backend.cli:main"

[tool.poetry]
packages = [{include = "vector_brain_backend"}]


# [project.dynamic]
# version = {attr = "vector_brain_backend.__version__"}
# readme = "README.md"  # Se for dinâmico

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
py_version = "311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=vector_brain_backend --cov-report=term-missing"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"